Fixed NSLog warning.

+++ the-blue-alliance-ios/Event+Create.m
@@ -23,7 +23,7 @@
     NSError *error;
     NSArray *matches = [context executeFetchRequest:fetchRequest error:&error];
     if (error || !matches || matches.count > 1) {
-        NSLog(@"ERROR searching for existing event with key %@. %lu matches found. Error: %@", key, matches.count, error);
+        NSLog(@"ERROR searching for existing event with key %@. %lu matches found. Error: %@", key, (unsigned long)matches.count, error);
     } else if (matches.count) {
         event = [matches firstObject];
     } else {
